Simplified the mouse that you don't need a voice command to turn it on: all you need to do is to start hissing
I remove the requirement that you need to know englishAnd you need to know the magic voice command to turn this mouse on.Now when you install these files in the Talon user directory, this mouse will start working as soon as you start hissing .

I also added a global variable and voice command that allows you to toggle whether the crosshairs should be shown all the time or only when you are hissing a.  This is not a full, official talent setting, but just a global variable.